auth

+++ src/app/auth/services/auth.service.ts
@@ -5,6 +5,28 @@ import { Router } from '@angular/router';
 import { Observable, Subscription } from 'rxjs';
 import { environment } from 'src/environments/environment';
 
+type LoginResponse = {
+  code: string,
+  data: {
+    token: string,
+    user: {
+      business: {
+        address: {
+          latitude: number,
+          longitude: number,
+        },
+        business_id: string,
+        business_name: string,
+        rating: number | null,
+      },
+      global_user_id: string,
+      email: string | null,
+      phone: string | null,
+      profile_picture_url: string | null,
+    }
+  }
+}
+
 @Injectable({
   providedIn: 'root'
 })
@@ -28,13 +50,30 @@ export class AuthService {
     if (user && token) {
       this.authData.isLoggedIn = true;
       this.authData.user = JSON.parse(user);
-      this.router.navigate(['/home']);
+      this.checkAuthorization(this.authData.user).subscribe({
+        next: (res) => {
+          console.log('checkAuthorization() res:', res);
+        },
+        error: (err) => {
+          console.log('checkAuthorization() err:', err);
+          this.logout();
+        }
+      })
     } else {
       this.authData.isLoggedIn = false;
       this.authData.user = null;
       this.router.navigate(['/login']);
-      this.logout();
+      this.logout('google-only');
     }
+  }
+
+  checkAuthorization(user: any) {
+    return this.http.get(`${environment.auth_endpoint}/check/${user.business_id}`, {
+      headers: {
+        'module': 'business',
+        'Authorization': `Bearer ${localStorage.getItem('token')}`
+      }
+    })
   }
 
   loginWithGoogle() {
@@ -46,19 +85,32 @@ export class AuthService {
   }
 
   checkBusiness(token: string) {
-    return this.http.post(environment.auth_endpoint + '/login', { token }, {
+    return this.http.post<LoginResponse>(environment.auth_endpoint + '/login', { token }, {
       headers: {
         'module': 'business'
       }
     })
   }
 
-  logout() {
-    this.authData.isLoggedIn = false;
-    localStorage.clear();
+  removeToken() {
+    return this.http.delete(`${environment.auth_endpoint}/logout`, {
+      headers: {
+        'module': 'business',
+        'Authorization': `Bearer ${localStorage.getItem('token')}`
+      }
+    })
+  }
+
+  logout(type: string = 'all') {
     this.auth.signOut().then((res) => {
       console.log(res);
-      this.router.navigate(['/login']);
+      if(type === 'google-only') return;
+      this.removeToken().subscribe((res) => {
+        localStorage.clear();
+        this.authData.user = null;
+        this.authData.isLoggedIn = false;
+        this.router.navigate(['/login']);
+      })
     });
   }
 }
